third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/script-charset-03-expected.txt

This is a testharness.js-based test.
FAIL Script changing @charset assert_equals: expected "śćążź" but got "ÂśĂ¦ÂąÂżÂź"
Harness: the test ran to completion.
